pyproject.toml
[project]
name = "cpc-camp-slack-bot"
version = "0.1.0"
description = "Research camp Slack bot that watches presentations and participates in discussions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "slack-bolt>=1.18",
    "slack-sdk>=3.27",
    "anthropic>=0.40",
    "pymupdf>=1.24",
    "pydantic-settings>=2.0",
    "aiohttp>=3.9",
    "python-frontmatter>=1.0",
    "sounddevice>=0.4",
    "faster-whisper>=1.0",
    "onnxruntime>=1.19,<1.24",
    "numpy>=1.26",
]

[project.scripts]
campbot = "campbot.main:cli"

[tool.uv]
package = true

[tool.uv.sources]
